feat: update docs to have the VisionModels, not sure why the save_pretrained method doesn't show up (#43)

+++ src/jimm/common/vit.py
@@ -180,7 +180,7 @@ class VisionTransformerBase(nnx.Module):
             raise ValueError("pooling_type must be either MAP or CLS.")
         self.position_embeddings = nnx.Param(pos_emb_value)
         vision_n_positions = n_patches + 1 if self.pooling_type == "CLS" else n_patches
-        self.vision_position_ids = nnx.Param(jnp.arange(vision_n_positions, dtype=jnp.int32).reshape(1, -1))
+        self.vision_position_ids = nnx.Param(jnp.arange(vision_n_positions, dtype=dtype).reshape(1, -1))
 
         if self.use_pre_norm:
             self.ln_pre = nnx.LayerNorm(
